refactor(app): extract helpers + buildRouter from run.go (#138)
Step 1 of the pkg/app/run.go split (1219 lines). Extracts two self-contained
units that have no coupling to the RunFree construction/registration flow:

- helpers.go (52 lines): serviceFunc wrapper + aiConfigFromConfig +
  webrtcICEServers converter. These are pure utilities used by RunFree.
- router.go (107 lines): buildRouter() — the chi HTTP router construction
  (middleware chain, API Key auth, metrics mount, WebDAV, upload, SPA static
  handler, cache headers). Already fully self-contained.

run.go shrinks from 1219 → 1082 lines (137 lines extracted). The RunFree
function body itself is kept intact in this PR because its construction phase
(steps 0-10) has deep cross-domain dependencies (camera manager needs merge +
transcode + rollingMerge + timelapseMergeMgr products; relay needs camera;
periodicMerge needs camMgr) that make a naive builder-file split increase
complexity rather than reduce it. The service registration order (db →
startup-bg → camera → health → ... → remoteLog) is load-bearing and must stay
in a single function.

A follow-up PR can extract the construction phase into an appBuilders
accumulator struct if the file size becomes a maintenance burden, but that
refactor carries goroutine-lifecycle risk (PR #144's startup-bg/api-handler/
mergeDone/cleanupDone joins) and is deferred until the construction dependency
graph stabilizes.

Zero behavior change. All tests pass. 6 imports removed from run.go
(io/fs, strings, chimiddleware, promhttp, ui, pionwebrtc) — they moved with
their code to helpers.go/router.go.
